extracted PropTypes to module
