+++ packages/inferno-compat/index.js
@@ -1,3 +1,2 @@
 module.exports = require('./dist').default;
 module.exports.default = module.exports;
-
