+++ packages/inferno-compat/prop-types.js
@@ -1,0 +1,2 @@
+module.exports = require('./dist/PropTypes').default;
+module.exports.default = module.exports;
