+++ packages/inferno-compat/src/index.ts
@@ -15,6 +15,7 @@ import infernoCreateElement from 'inferno-create-element';
 import { isArray, isBrowser, isFunction, isNull, isNullOrUndef, isString, NO_OP } from 'inferno-shared';
 import _VNodeFlags from 'inferno-vnode-flags';
 import isValidElement from './isValidElement';
+import PropTypes from './PropTypes';
 import SVGDOMPropertyConfig from './SVGDOMPropertyConfig';
 
 declare global {
@@ -22,32 +23,6 @@ declare global {
 		persist: Function;
 	}
 }
-
-// Inlined PropTypes, there is propType checking ATM.
-// tslint:disable-next-line:no-empty
-function proptype() {}
-(proptype as any).isRequired = proptype;
-
-const getProptype = () => proptype;
-
-const PropTypes = {
-	any: getProptype,
-	array: proptype,
-	arrayOf: getProptype,
-	bool: proptype,
-	element: getProptype,
-	func: proptype,
-	instanceOf: getProptype,
-	node: getProptype,
-	number: proptype,
-	object: proptype,
-	objectOf: getProptype,
-	oneOf: getProptype,
-	oneOfType: getProptype,
-	shape: getProptype,
-	string: proptype,
-	symbol: proptype
-};
 
 options.findDOMNodeEnabled = true;
